Moved all attribute drawers to the editor folder
