+++ Assets/Etra Games/Etra'sStarterAssets/1-UserAssets/CharacterAsset/Scripts/Abilities/FPS_ExclusiveAbilities/ABILITY_FPS_Crouch.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 namespace Etra.StarterAssets.Abilities.FirstPerson
 {
     [RequireComponent(typeof(ABILITY_CharacterMovement))]
-    [AbilityUsage(EtraCharacterMainController.GameplayTypeFlags.FirstPerson)]
+    [AbilityUsageAttribute(EtraCharacterMainController.GameplayTypeFlags.FirstPerson)]
     public class ABILITY_FPS_Crouch : EtraAbilityBaseClass
     {
         //This script is functional and as is for now. Will rework in the future.
