+++ Assets/Etra Games/Etra'sStarterAssets/1-UserAssets/CharacterAsset/Scripts/Abilities/General_Abilities/ABILITY_GrabInteract.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 namespace Etra.StarterAssets.Abilities
 {
-    [AbilityUsage(EtraCharacterMainController.GameplayTypeFlags.All)]
+    [AbilityUsageAttribute(EtraCharacterMainController.GameplayTypeFlags.All)]
     public class ABILITY_GrabInteract : EtraAbilityBaseClass
     {
         //Abilty created by: asour
