+++ Assets/Etra Games/Etra'sStarterAssets/1-UserAssets/CharacterAsset/Scripts/Abilities/General_Abilities/ABILITY_Sprint.cs
@@ -4,7 +4,7 @@ namespace Etra.StarterAssets.Abilities
 {
     [RequireComponent(typeof(ABILITY_CharacterMovement))]
 
-    [AbilityUsage(EtraCharacterMainController.GameplayTypeFlags.All)]
+    [AbilityUsageAttribute(EtraCharacterMainController.GameplayTypeFlags.All)]
     public class ABILITY_Sprint : EtraAbilityBaseClass
     {
         //This script is just a contaiiner for the sprint speed variable.
