+++ Assets/Etra Games/Etra'sStarterAssets/2-Source/Editor/Attributes.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6b57db507cb2f9b4d9674974a01133f2
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
