+++ Assets/Etra Games/Etra'sStarterAssets/2-Source/Editor/Attributes/ListToPopupDrawer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a79755dbb89e9b147bab27f6759c4299
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
