+++ Assets/Etra Games/Etra'sStarterAssets/2-Source/Scripts/Attributes/AbilityUsageAttribute.cs.meta
